修改了显示投注信息的问题 	modified:   res/layout/arrangefive_switchtabs_activitygroup.xml 	modified:   res/layout/arrangethree_switchtabs_activitygroup.xml 	modified:   res/layout/betbar_included.xml 	modified:   res/layout/bettingbar.xml 	modified:   res/layout/constantly_switchtabs_activitygroup.xml 	modified:   res/layout/doubleball_switchtabs_activitygroup.xml 	modified:   res/layout/elevenluckgold_spinners_activitygroup.xml 	modified:   res/layout/guangdonghappyvery_spinners_activitygroup.xml 	modified:   res/layout/jiangxielevenselectfive_spinners_activitygroup.xml 	modified:   res/layout/servenhappy_switchtabs_activitygroup.xml 	modified:   res/layout/servenstar_switchtabs_activitygroup.xml 	modified:   res/layout/spinners_activitygroup.xml 	modified:   res/layout/superlotto_switchtabs_activitygroup.xml 	modified:   res/layout/twentytwoselectfive_switchtabs_activitygroup.xml 	modified:   res/layout/welfare3d_switchtabs_activitygroup.xml 	modified:   src/com/ruyicai/android/controller/activity/home/buylotteryhall/LotterySwitchTabsActivityGroup.java 	modified:   src/com/ruyicai/android/controller/activity/home/buylotteryhall/SwitchTabsActivityGroup.java 	modified:   src/com/ruyicai/android/controller/activity/home/buylotteryhall/doubleball/DoubleBallSwitchTabsActivityGroup.java 	renamed:    src/com/ruyicai/android/controller/compontent/bar/BetBar.java -> src/com/ruyicai/android/controller/compontent/bar/BettingBar.java 	modified:   src/com/ruyicai/android/controller/compontent/dialog/DialogType.java 	renamed:    src/com/ruyicai/android/model/bean/BettingInfo.java -> src/com/ruyicai/android/model/bean/BettingInformation.java 	modified:   src/com/ruyicai/android/model/bean/NumberBasket.java 	modified:   src/com/ruyicai/android/model/bean/lottery/DoubleBallLottery.java 	modified:   src/com/ruyicai/android/tools/MathTools.java

+++ src/com/ruyicai/android/model/bean/lottery/DoubleBallLottery.java
@@ -207,8 +207,8 @@ public class DoubleBallLottery extends Lottery {
 		int redNum = aNowSelectedNumberLists.get(0).size();
 		int blueNum = aNowSelectedNumberLists.get(1).size();
 
-		int betNumber = MathTools.combination(redNum, 6) * MathTools.combination(blueNum, 1);
-		int betAmount = betNumber * 2;
+		long betNumber = MathTools.combination(redNum, 6) * MathTools.combination(blueNum, 1);
+		long betAmount = betNumber * 2;
 
 		StringBuilder betInfStringBuilder = new StringBuilder();
 		betInfStringBuilder.append("共").append(betNumber).append("注").append(betAmount).append("元");
